+++ tests/hello_world_early.c
@@ -78,7 +78,7 @@ int alice_status(enum nmp_status s, const union nmp_cb_status *d, void *c)
 }
 
 
-void bob_data(const u8 *data, const u16 len, void *drv)
+void bob_data(const u8 *data, const u32 len, void *drv)
 {
         if (len != (u16) strlen(msg))
                 test_fail();
